资产清单-按照连接协议来过滤资产异常

See merge request codedriver/codedriver-cmdb-base!70

+++ src/main/java/codedriver/framework/cmdb/dto/resourcecenter/ResourceSearchVo.java
@@ -21,7 +21,7 @@ public class ResourceSearchVo extends BasePageVo {
     private String port;
     private Long envId;
     private Long appModuleId;
-    private List<Integer> protocolIdList;
+    private List<Long> protocolIdList;
     private List<String> protocolList;
     private List<Long> stateIdList;
     private List<Long> envIdList;
@@ -91,11 +91,11 @@ public class ResourceSearchVo extends BasePageVo {
         this.appModuleId = appModuleId;
     }
 
-    public List<Integer> getProtocolIdList() {
+    public List<Long> getProtocolIdList() {
         return protocolIdList;
     }
 
-    public void setProtocolIdList(List<Integer> protocolIdList) {
+    public void setProtocolIdList(List<Long> protocolIdList) {
         this.protocolIdList = protocolIdList;
     }
